refactor: tool bar items files and add subtitle in oudsNavigationTitle helper
