+++ OUDS/Core/Components/Sources/Navigations/ToolBar/OUDSToolBarBottom.swift
@@ -20,10 +20,6 @@ import SwiftUI
 ///
 /// ``oudsToolBarBottom`` View helper applies a SwiftUI toolBar configuration.
 ///
-/// ## Appearances
-///
-/// TODO: #1174 - Update this section
-///
 /// ## Platform considerations
 ///
 /// - Available on iOS 15, macOS 15, visionOS 1
@@ -82,7 +78,7 @@ import SwiftUI
 /// ![A bottom toolBar component in light mode with Liquid Glass effect and Wireframe theme](component_toolBarBottom_Wireframe_light)
 /// ![A bottom toolBar component in dark mode with Liquid Glass effect and Wireframe theme](component_toolBarBottom_Wireframe_dark)
 ///
-/// - Since: 1.3.0
+/// - Since: 1.4.0
 extension View {
 
     /// Creates a bottom toolBar with leading and trailing items.
@@ -90,12 +86,14 @@ extension View {
     /// - Parameters:
     ///   - leadingItems: The items displayed on the leading side.
     ///   - trailingItems: The items displayed on the trailing side.
+    ///
+    /// - Since: 1.4.0
     @available(iOS 15, macOS 13, visionOS 1, *)
     public func oudsToolBarBottom(@OUDSToolBarItemsBuilder leadingItems: () -> [OUDSToolBarItem] = { [] },
                                   @OUDSToolBarItemsBuilder trailingItems: () -> [OUDSToolBarItem] = { [] }) -> some View
     {
-        modifier(OUDSToolBarBottomModifier(leadingItems: leadingItems,
-                                           trailingItems: trailingItems))
+        modifier(ToolBarBottomModifier(leadingItems: leadingItems,
+                                       trailingItems: trailingItems))
     }
 
     /// Creates a bottom toolBar with grouped items, centered to the screen.
@@ -103,92 +101,11 @@ extension View {
     /// **Warning: Works only with iOS 26+ / Liquid Glass, otherwise items will be splitted by the system**
     ///
     /// - Parameter groupedItems: All the items to place in the center of the s creen
+    ///
+    /// - Since: 1.4.0
     @available(iOS 15, macOS 13, visionOS 1, *)
     public func oudsToolBarBottom(@OUDSToolBarItemsBuilder groupedItems: () -> [OUDSToolBarItem] = { [] }) -> some View {
-        modifier(OUDSToolBarBottomModifier(groupedItems: groupedItems))
+        modifier(ToolBarBottomModifier(groupedItems: groupedItems))
     }
 }
-
-@available(iOS 15, macOS 13, visionOS 1, *)
-struct OUDSToolBarBottomModifier: ViewModifier {
-
-    // MARK: - Stored properties
-
-    private let leadingItems: [OUDSToolBarItem]?
-    private let trailingItems: [OUDSToolBarItem]?
-    private let groupedItems: [OUDSToolBarItem]?
-
-    @Environment(\.theme) private var theme
-
-    // MARK: - Initializer
-
-    /// Creates a bottom toolBar with leading and trailing items, and a space between them.
-    ///
-    /// - Parameters:
-    ///   - leadingItems: The items displayed on the leading side.
-    ///   - trailingItems: The items displayed on the trailing side.
-    ///   - content: The content view wrapped by the toolBar.
-    init(@OUDSToolBarItemsBuilder leadingItems: () -> [OUDSToolBarItem] = { [] },
-         @OUDSToolBarItemsBuilder trailingItems: () -> [OUDSToolBarItem] = { [] })
-    {
-        groupedItems = nil
-        self.leadingItems = leadingItems()
-        self.trailingItems = trailingItems()
-    }
-
-    init(@OUDSToolBarItemsBuilder groupedItems: () -> [OUDSToolBarItem] = { [] }) {
-        self.groupedItems = groupedItems()
-        leadingItems = nil
-        trailingItems = nil
-    }
-
-    /// Creates a bottom toolBar with grouped items, centered to the screen.
-    ///
-    /// **Warning: Works only with iOS 26+ / Liquid Glass, otherwise items will be splitted by the system**
-    ///
-    /// - Parameters:
-    ///   - groupedItems: All the items to place in the center of the s creen
-    ///   - content: The content view wrapped by the toolBar.
-    func body(content: Content) -> some View {
-        if let groupedItems {
-            content
-                .toolbar {
-                    ToolbarItemGroup(placement: bottomPlacement) {
-                        itemsView(groupedItems)
-                    }
-                }
-        } else {
-            content
-                .toolbar {
-                    ToolbarItemGroup(placement: bottomPlacement) {
-                        itemsView(leadingItems)
-                        Spacer()
-                        itemsView(trailingItems)
-                    }
-                }
-        }
-    }
-
-    // MARK: - Helpers
-
-    private var bottomPlacement: ToolbarItemPlacement {
-        #if os(iOS) || os(visionOS)
-        return .bottomBar
-        #elseif os(macOS)
-        return .automatic
-        #else
-        return .automatic
-        #endif
-    }
-
-    @ViewBuilder
-    private func itemsView(_ items: [OUDSToolBarItem]?) -> some View {
-        if let items, !items.isEmpty {
-            ForEach(items) { item in
-                item.modifier(ToolBarBottomItemActionStyleModifier())
-            }
-        }
-    }
-}
-
 #endif
